keep segment as uint16 for blockmap, build blockmap progressively
git-svn-id: https://svn.osgeo.org/gdal@20252 f0d54148-0727-0410-94bb-9a71ac55c965

+++ sandbox/warmerdam/pcidsk/src/segment/vecsegdataindex.cpp
@@ -274,7 +274,7 @@ void VecSegDataIndex::VacateBlockRange( uint32 start, uint32 count )
     GetIndex(); // make sure loaded.
 
     unsigned int i;
-    uint32  next_block = vs->GetContentSize() / block_page_size;
+    uint32  next_block = (uint32) (vs->GetContentSize() / block_page_size);
 
     for( i = 0; i < block_count; i++ )
     {
